requirements.txt
biopython
openai==1.13.3
python-dotenv==1.0.0
streamlit==1.31.0
requests==2.31.0 